+++ src/launcher/hook/create_watchdog.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <autoattach/lua_module.h>
+#include <memory>
+
+namespace luadebug::lua {
+    struct resolver;
+}
+
+namespace luadebug::autoattach {
+    struct watchdog;
+    watchdog* create_watchdog(lua_version v, const lua::resolver& resolver);
+}
